Sanitize default filenames when saving

+++ tests/test_genius.py
@@ -7,7 +7,7 @@ from lyricsgenius.artist import Artist
 # Import client access token from environment variable
 client_access_token = os.environ.get("GENIUS_CLIENT_ACCESS_TOKEN", None)
 assert client_access_token is not None, "Must declare environment variable: GENIUS_CLIENT_ACCESS_TOKEN"
-api = Genius(client_access_token, sleep_time=0.5)
+genius = Genius(client_access_token, sleep_time=0.5)
 
 
 class TestEndpoints(unittest.TestCase):
@@ -20,7 +20,7 @@ class TestEndpoints(unittest.TestCase):
     def test_search_genius_web(self):
         # TODO: test more than just a 200 response
         msg = "Response was None."
-        r = api.search_genius_web(self.search_term)
+        r = genius.search_genius_web(self.search_term)
         self.assertTrue(r is not None, msg)
 
 
@@ -32,8 +32,13 @@ class TestArtist(unittest.TestCase):
         cls.artist_name = "The Beatles"
         cls.new_song = "Paperback Writer"
         cls.max_songs = 2
-        cls.artist = api.search_artist(
+        cls.artist = genius.search_artist(
             cls.artist_name, max_songs=cls.max_songs)
+
+    def sanitize_filename(f):
+        keepchars = (" ", ".", "_")
+        s = "".join(c for c in f if c.isalnum() or c in keepchars).rstrip()
+        return s
 
     def test_artist(self):
         msg = "The returned object is not an instance of the Artist class."
@@ -45,19 +50,29 @@ class TestArtist(unittest.TestCase):
 
     def test_add_song_from_same_artist(self):
         msg = "The new song was not added to the artist object."
-        self.artist.add_song(api.search_song(self.new_song, self.artist_name))
+        self.artist.add_song(genius.search_song(self.new_song, self.artist_name))
         self.assertEqual(self.artist.num_songs, self.max_songs+1, msg)
 
     def test_add_song_from_different_artist(self):
         msg = "A song from a different artist was incorrectly allowed to be added."
-        self.artist.add_song(api.search_song("These Days", "Jackson Browne"))
+        self.artist.add_song(genius.search_song("These Days", "Jackson Browne"))
         self.assertEqual(self.artist.num_songs, self.max_songs, msg)
+
+    def determine_filenames(self, extension):
+        expected_filenames = []
+        for song in self.artist.songs:
+            fn = "lyrics_{name}_{song}.{ext}".format(name=self.artist.name,
+                                                     song=song.title,
+                                                     ext=extension)
+            fn = song.sanitize_filename(fn.lower().replace(" ", ""))
+            expected_filenames.append(fn)
+        return expected_filenames
 
     def test_saving_json_file(self):
         print('\n')
-        format_ = 'json'
-        msg = "Could not save {} file.".format(format_)
-        expected_filenames = [('lyrics_thebeatles_{}.'.format(song.title.replace(" ", "").lower()) + format_) for song in self.artist.songs]
+        extension = 'json'
+        msg = "Could not save {} file.".format(extension)
+        expected_filenames = self.determine_filenames(extension)
 
         # Remove the test file if it already exists
         for expected_filename in expected_filenames:
@@ -65,26 +80,23 @@ class TestArtist(unittest.TestCase):
                 os.remove(expected_filename)
 
         # Test saving json file
-        self.artist.save_lyrics(extension=format_, overwrite=True)
+        self.artist.save_lyrics(extension=extension, overwrite=True)
         for expected_filename in expected_filenames:
             self.assertTrue(os.path.isfile(expected_filename), msg)
 
-        # Test overwriting json file
+        # Test overwriting json file (now that file is written)
         try:
-            self.artist.save_lyrics(
-                extension=format_, overwrite=True)
-            for expected_filename in expected_filenames:
-                os.remove(expected_filename)
+            self.artist.save_lyrics(extension=extension, overwrite=True)
         except:
-            self.fail("Failed {} overwrite test".format(format_))
-            for expected_filename in expected_filenames:
-                os.remove(expected_filename)
+            self.fail("Failed {} overwrite test".format(extension))
+        for expected_filename in expected_filenames:
+            os.remove(expected_filename)
 
     def test_saving_txt_file(self):
         print('\n')
-        format_ = 'txt'
-        msg = "Could not save {} file.".format(format_)
-        expected_filenames = [('lyrics_thebeatles_{}.'.format(song.title.replace(" ", "").lower()) + format_) for song in self.artist.songs]
+        extension = 'txt'
+        msg = "Could not save {} file.".format(extension)
+        expected_filenames = self.determine_filenames(extension)
 
         # Remove the test file if it already exists
         for expected_filename in expected_filenames:
@@ -92,20 +104,18 @@ class TestArtist(unittest.TestCase):
                 os.remove(expected_filename)
 
         # Test saving txt file
-        self.artist.save_lyrics(extension=format_, overwrite=True)
+        self.artist.save_lyrics(extension=extension, overwrite=True)
         for expected_filename in expected_filenames:
             self.assertTrue(os.path.isfile(expected_filename), msg)
 
-        # Test overwriting txt file
+        # Test overwriting txt file (now that file is written)
         try:
             self.artist.save_lyrics(
-                extension=format_, overwrite=True)
-            for expected_filename in expected_filenames:
-                os.remove(expected_filename)
+                extension=extension, overwrite=True)
         except:
-            self.fail("Failed {} overwrite test".format(format_))
-            for expected_filename in expected_filenames:
-                os.remove(expected_filename)
+            self.fail("Failed {} overwrite test".format(extension))
+        for expected_filename in expected_filenames:
+            os.remove(expected_filename)
 
 
 class TestSong(unittest.TestCase):
@@ -117,9 +127,9 @@ class TestSong(unittest.TestCase):
         cls.song_title = 'begin again'  # Lowercase is intentional
         cls.album = 'The Party'
         cls.year = '2016-05-20'
-        cls.song = api.search_song(cls.song_title, cls.artist_name)
-        api.remove_section_headers = True
-        cls.song_trimmed = api.search_song(cls.song_title, cls.artist_name)
+        cls.song = genius.search_song(cls.song_title, cls.artist_name)
+        genius.remove_section_headers = True
+        cls.song_trimmed = genius.search_song(cls.song_title, cls.artist_name)
 
     def test_song(self):
         msg = "The returned object is not an instance of the Song class."
@@ -127,8 +137,8 @@ class TestSong(unittest.TestCase):
 
     def test_title(self):
         msg = "The returned song title does not match the title of the requested song."
-        self.assertEqual(api._clean_str(self.song.title),
-                         api._clean_str(self.song_title), msg)
+        self.assertEqual(genius._clean_str(self.song.title),
+                         genius._clean_str(self.song_title), msg)
 
     def test_artist(self):
         msg = "The returned artist name does not match the artist of the requested song."
@@ -156,13 +166,13 @@ class TestSong(unittest.TestCase):
 
     def test_result_is_lyrics(self):
         msg = "Did not reject a false-song."
-        self.assertFalse(api._result_is_lyrics('Beatles Tracklist'), msg)
+        self.assertFalse(genius._result_is_lyrics('Beatles Tracklist'), msg)
 
     def test_saving_json_file(self):
         print('\n')
-        format_ = 'json'
-        msg = "Could not save {} file.".format(format_)
-        expected_filename = 'lyrics_save_test_file.' + format_
+        extension = 'json'
+        msg = "Could not save {} file.".format(extension)
+        expected_filename = 'lyrics_save_test_file.' + extension
         filename = expected_filename.split('.')[0]
 
         # Remove the test file if it already exists
@@ -170,23 +180,23 @@ class TestSong(unittest.TestCase):
             os.remove(expected_filename)
 
         # Test saving json file
-        self.song.save_lyrics(filename=filename, extension=format_, overwrite=True)
+        self.song.save_lyrics(filename=filename, extension=extension, overwrite=True)
         self.assertTrue(os.path.isfile(expected_filename), msg)
 
-        # Test overwriting json file
+        # Test overwriting json file (now that file is written)
         try:
             self.song.save_lyrics(
-                filename=expected_filename, extension=format_, overwrite=True)
+                filename=expected_filename, extension=extension, overwrite=True)
             os.remove(expected_filename)
         except:
-            self.fail("Failed {} overwrite test".format(format_))
+            self.fail("Failed {} overwrite test".format(extension))
             os.remove(expected_filename)
 
     def test_saving_txt_file(self):
         print('\n')
-        format_ = 'txt'
-        msg = "Could not save {} file.".format(format_)
-        expected_filename = 'lyrics_save_test_file.' + format_
+        extension = 'txt'
+        msg = "Could not save {} file.".format(extension)
+        expected_filename = 'lyrics_save_test_file.' + extension
         filename = expected_filename.split('.')[0]
 
         # Remove the test file if it already exists
@@ -194,14 +204,32 @@ class TestSong(unittest.TestCase):
             os.remove(expected_filename)
 
         # Test saving txt file
-        self.song.save_lyrics(filename=filename, extension=format_, overwrite=True)
+        self.song.save_lyrics(filename=filename,
+                              extension=extension,
+                              overwrite=True)
         self.assertTrue(os.path.isfile(expected_filename), msg)
 
-        # Test overwriting txt file
+        # Test overwriting txt file (now that file is written)
         try:
-            self.song.save_lyrics(
-                filename=filename, extension=format_, overwrite=True)
-            os.remove(expected_filename)
+            self.song.save_lyrics(filename=filename,
+                                  extension=extension,
+                                  overwrite=True)
         except:
-            self.fail("Failed {} overwrite test".format(format_))
+            self.fail("Failed {} overwrite test".format(extension))
+        os.remove(expected_filename)
+
+    def test_bad_chars_in_filename(self):
+        print("\n")
+        extension = "json"
+        msg = "Could not save {} file.".format(extension)
+        song = genius.search_song("Blessed rainbow", "Ariana Grande")
+        expected_filename = "lyrics_arianagrande_blessedrainbow.json"
+
+        # Remove the test file if it already exists
+        if os.path.isfile(expected_filename):
             os.remove(expected_filename)
+
+        # Test saving txt file
+        song.save_lyrics(extension=extension, overwrite=True)
+        self.assertTrue(os.path.isfile(expected_filename), msg)
+        os.remove(expected_filename)
